03/08/2026

+++ backend/app/auth/services/redis_service.py
@@ -4,44 +4,16 @@ from app.database import redis
 class RedisService:
 
     @staticmethod
-    async def save_refresh_token(
+    async def save_refresh_token(user_id,token):
 
-        student_id,
-
-        token
-
-    ):
-
-        await redis.redis_client.set(
-
-            f"refresh:{student_id}",
-
-            token
-
-        )
+        await redis.redis_client.set(f"refresh:{user_id}",token)
 
     @staticmethod
-    async def get_refresh_token(
+    async def get_refresh_token(user_id):
 
-        student_id
-
-    ):
-
-        return await redis.redis_client.get(
-
-            f"refresh:{student_id}"
-
-        )
+        return await redis.redis_client.get(f"refresh:{user_id}")
 
     @staticmethod
-    async def delete_refresh_token(
+    async def delete_refresh_token(user_id):
 
-        student_id
-
-    ):
-
-        await redis.redis_client.delete(
-
-            f"refresh:{student_id}"
-
-        )
+        await redis.redis_client.delete(f"refresh:{user_id}")
